webhook: Move httpapi into own package
